Move test to new security_headers_test
